Revert "Add cache user feature for autologin (#7775)"
This reverts commit 046557da9b5680c4ac6171dd32a0eec2ee6c358f.
