+++ unlock-app/src/components/interface/Authenticate.jsx
@@ -23,7 +23,6 @@ import { ConfigContext } from '../../utils/withConfig'
 import UnlockPropTypes from '../../propTypes'
 
 import LogInSignUp from './LogInSignUp'
-import { useAutoLogin } from '../../hooks/useAutoLogin'
 
 const GraphServiceProvider = GraphServiceContext.Provider
 
@@ -122,12 +121,6 @@ export const Authenticate = ({
     disconnectProvider()
   }
 
-  const { tryAutoLogin, isLoading } = useAutoLogin({ authenticate })
-
-  useEffect(() => {
-    tryAutoLogin()
-  }, [])
-
   return (
     <AuthenticationContext.Provider
       value={{
@@ -153,7 +146,7 @@ export const Authenticate = ({
             email={email}
             encryptedPrivateKey={encryptedPrivateKey}
           >
-            {!isLoading ? children : <Loading />}
+            {children}
           </Providers>
         </WalletServiceContext.Provider>
       )}
